// src/app/shared/validators/enums/errorAssociation.enum.ts
export enum ErrorAssociation {
  REQUIRED = 'required',
  MINLENGTH = 'minlength',
  MISMATCH = 'misMatch',
  EMAIL = 'email',
  REQUIRES_UPPERCASE = 'requiresUppercase',
  REQUIRES_LOWERCASE = 'requiresLowercase',
  REQUIRES_DIGIT = 'requiresDigit',
  REQUIRES_SPECIAL_CHARS = 'requiresSpecialChars',
}